2.0.0/requirements.txt
requests>=2.31.0,<3.0.0
beautifulsoup4>=4.12.0,<5.0.0
PyPDF2>=3.0.0,<4.0.0
langchain>=0.2.0,<0.3.0
sentence-transformers>=2.2.0,<3.0.0
numpy>=1.24.0,<2.0.0
--extra-index-url https://download.pytorch.org/whl/cu118
torch
faiss-gpu